Initial neural networks prediction.
Try the classification of three different outcomes: underperform,
neutral, overperform, seeing the last days of hourly prices and trying
to predict if the target price of the next hours will be in one of the
three classes. The neural net outperforms a lot the error you have just
picking the most common outcome in the testing set, however this is
just the result of using interleaved data, so it "learns" that a given
curve can be seen shifted.