NN: code consolidation, but it does not work.
As expected the NN is not able to predict anything, since the stock
market is very hard to predict from the timeseries of its prices: you
miss all the informations that really change the stock price. Yet I
hoped that in the short period there was some pattern: no way. But it
was funny to try.