Case insensitive lists and stocks names.
In order to be backward compatible, we use case insensitive collation
both in the creation of the tables and in the queries. Old users will
still have case sensitive tables already created. Moreover it is not
possible to do an SQLite alter table to change the table from case
sensitive to case insensitive.